+++ web/app/lukio/LukionOppiaineEditor.jsx
@@ -1,16 +1,20 @@
-import React from 'react'
+import React from 'baret'
+import R from 'ramda'
 
 import {t} from '../i18n/i18n'
-import {modelData, modelItems} from '../editor/EditorModel.js'
+import {modelData, modelItems, modelLookup, modelTitle, pushRemoval} from '../editor/EditorModel.js'
 import {suorituksenTilaSymbol} from '../suoritus/Suoritustaulukko'
 import {KurssitEditor} from '../kurssi/KurssitEditor'
 import {tilaText} from '../suoritus/Suoritus'
 import {FootnoteHint} from '../components/footnote'
-import {modelLookup, modelTitle, pushRemoval} from '../editor/EditorModel'
 import {isKieliaine, isLukionMatematiikka, isPaikallinen} from '../suoritus/Koulutusmoduuli'
 import {Editor} from '../editor/Editor'
 import {ArvosanaEditor} from '../suoritus/ArvosanaEditor'
 import {PropertiesEditor} from '../editor/PropertiesEditor'
+import {saveOrganizationalPreference} from '../virkailija/organizationalPreferences'
+import {paikallinenOppiainePrototype} from '../perusopetus/PerusopetuksenOppiaineEditor'
+import {doActionWhileMounted} from '../util/util'
+import {createOppiaineenSuoritus} from './lukio'
 
 const Nimi = ({oppiaine}) => {
   const {edit} = oppiaine.context
@@ -72,36 +76,65 @@ const Arviointi = ({oppiaine, suoritetutKurssit, footnote}) => {
   )
 }
 
-export const LukionOppiaineRowEditor = ({oppiaine, footnote, allowOppiaineRemoval = true}) => {
-  const kurssit = modelItems(oppiaine, 'osasuoritukset')
-  const suoritetutKurssit = kurssit.map(k => modelData(k)).filter(k => k.arviointi)
-  const {edit} = oppiaine.context
+export class LukionOppiaineRowEditor extends React.Component {
+  saveChangedPreferences() {
+    const {oppiaine} = this.props
 
-  return (
-    <tr className={'oppiaine oppiaine-rivi ' + modelData(oppiaine, 'koulutusmoduuli.tunniste.koodiarvo')}>
-      <td className='suorituksentila' title={tilaText(oppiaine)}>
-        <div>
-          {suorituksenTilaSymbol(oppiaine)}
-        </div>
-      </td>
-      <td className='oppiaine'>
-        <div className='title'>
-          <Nimi oppiaine={oppiaine}/>
-          <KoulutusmoduuliPropertiesEditor oppiaine={oppiaine}/>
-        </div>
-        <KurssitEditor model={oppiaine}/>
-      </td>
-      <td className='maara'>{suoritetutKurssit.length}</td>
-      <td className='arvosana'>
-        <Arviointi oppiaine={oppiaine} suoritetutKurssit={suoritetutKurssit} footnote={footnote}/>
-      </td>
-      {
-        edit && allowOppiaineRemoval && (
-          <td className='remove-row'>
-            <a className='remove-value' onClick={() => pushRemoval(oppiaine)}/>
-          </td>
-        )
-      }
-    </tr>
-  )
+    const data = modelData(oppiaine, 'koulutusmoduuli')
+    const organisaatioOid = modelData(oppiaine.context.toimipiste).oid
+    const key = data.tunniste.koodiarvo
+
+    saveOrganizationalPreference(
+      organisaatioOid,
+      paikallinenOppiainePrototype(createOppiaineenSuoritus(oppiaine.context.suoritus)).value.classes[0],
+      key,
+      data
+    )
+  }
+
+  render() {
+    const {oppiaine, footnote, allowOppiaineRemoval = true} = this.props
+    const kurssit = modelItems(oppiaine, 'osasuoritukset')
+    const suoritetutKurssit = kurssit.map(k => modelData(k)).filter(k => k.arviointi)
+    const {edit} = oppiaine.context
+
+    return (
+      <tr className={'oppiaine oppiaine-rivi ' + modelData(oppiaine, 'koulutusmoduuli.tunniste.koodiarvo')}>
+        <td className='suorituksentila' title={tilaText(oppiaine)}>
+          <div>
+            {suorituksenTilaSymbol(oppiaine)}
+          </div>
+        </td>
+        <td className='oppiaine'>
+          <div className='title'>
+            <Nimi oppiaine={oppiaine}/>
+            <KoulutusmoduuliPropertiesEditor oppiaine={oppiaine}/>
+          </div>
+          <KurssitEditor model={oppiaine}/>
+        </td>
+        <td className='maara'>{suoritetutKurssit.length}</td>
+        <td className='arvosana'>
+          <Arviointi oppiaine={oppiaine} suoritetutKurssit={suoritetutKurssit} footnote={footnote}/>
+        </td>
+        {
+          edit && allowOppiaineRemoval && (
+            <td className='remove-row'>
+              <a className='remove-value' onClick={() => pushRemoval(oppiaine)}/>
+            </td>
+          )
+        }
+        {
+          this.state && this.state.changed && isPaikallinen(modelLookup(oppiaine, 'koulutusmoduuli')) &&
+          doActionWhileMounted(oppiaine.context.saveChangesBus, this.saveChangedPreferences.bind(this))
+        }
+      </tr>
+    )
+  }
+
+  componentWillReceiveProps(nextProps) {
+    const currentData = modelData(this.props.oppiaine)
+    const newData = modelData(nextProps.oppiaine)
+
+    if (!R.equals(currentData, newData)) this.setState({changed: true})
+  }
 }
