+++ web/app/lukio/lukio.js
@@ -1,3 +1,5 @@
+import {wrapOptional, modelLookup, modelItems, contextualizeSubModel, oneOfPrototypes} from '../editor/EditorModel'
+
 const perusteenDiaarinumeroToOppimäärä = diaarinumero => {
   switch (diaarinumero) {
     case '60/011/2015':
@@ -9,6 +11,16 @@ const perusteenDiaarinumeroToOppimäärä = diaarinumero => {
   }
 }
 
+const createOppiaineenSuoritus = (model, suoritusClass) => {
+  const oppiaineet = wrapOptional(modelLookup(model, 'osasuoritukset'))
+  const newItemIndex = modelItems(oppiaineet).length
+  const oppiaineenSuoritusProto = contextualizeSubModel(oppiaineet.arrayPrototype, oppiaineet, newItemIndex)
+  const options = oneOfPrototypes(oppiaineenSuoritusProto)
+  const proto = suoritusClass && options.find(p => p.value.classes.includes(suoritusClass)) || options[0]
+  return contextualizeSubModel(proto, oppiaineet, newItemIndex)
+}
+
 export {
-  perusteenDiaarinumeroToOppimäärä
+  perusteenDiaarinumeroToOppimäärä,
+  createOppiaineenSuoritus
 }
